Provide the gallery access cookie when downloading
At some point it started requiring the cookie when downloading images.
Or at the very least, it changed behaviour in such a way where providing
the cookie makes it work again.

To be able to provide the cookie when downloading, we thus now have to store
it in the JSON file. Previously generated JSON files can no longer be used,
and we break compatibility with Wipie's downloader. Regrettable.

Closes #3.